Move RestartManager related code into common
