+++ src/common/RestartManagement.h
@@ -1,0 +1,5 @@
+#pragma once
+
+#include <string>
+
+void RestartProcess(const std::wstring& processName);
